[CLI] Move SKILL.md generation into _skills and fold populate helper
Move build_skill_md() and its rendering helpers from skills.py into
_skills.py so _install_generated_skill imports it at module top instead
of a function-level import; skills.py now only holds the CLI commands.
The only deferred import left is .hf.app inside build_skill_md, which is
unavoidable (hf.py aggregates the command modules into the app walked).

Also fold _populate_install_dir into a populate closure in
_install_marketplace_skill, symmetric with _install_generated_skill.
